requirements.txt
streamlit
mysql-connector-python
scikit-learn
joblib
nltk
numpy
requests # Make sure this is present or add it